Add config_defaults hash parameter
Allow sites to specify a baseline set of config items to be applied to
all nodes via config_defaults that can be overridden via hash
merge with config_hash. This doesn't change the current behavior
for users, but allows users running both agents and servers
to consolidate defaults for both roles in a single key instead
of repeating the defaults merely to change a small number of config
options such as server: true and bootstrap_expect on server nodes.